GCG-13: As a user, I want to search friends by GitHub username so that I can connect.
refactor: move user search combobox into separate component
